Blog-Generator-App-with-Claude-API/requirements.txt
anthropic 
python-dotenv 
streamlit 